Reorganized getting started page into separate sections with side navbar
